@@ -1,8 +1,8 @@
 package com.view.recycler.floats;
 
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.yline.base.BaseAppCompatActivity;
 import com.yline.view.recycler.adapter.AbstractCommonRecyclerAdapter;
