@@ -1,8 +1,8 @@
 package com.recycler.snap.snap;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class PagerSnapCallbackHelper {
     private OnItemSelectListener onItemSelectListener;
